requirements.txt
blinker==1.8.2
click==8.1.7
flask==3.0.3
Flask-Cors==4.0.1
greenlet==3.0.3
gunicorn==22.0.0
importlib-metadata==7.2.1
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
mysqlclient==2.2.4
packaging==24.1
python-dotenv==1.0.1
SQLAlchemy==2.0.31
typing-extensions==4.12.2
werkzeug==3.0.3
zipp==3.19.2
